requirements.txt
numpy==1.23.4
opencv_contrib_python==4.5.5.62
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.5.0
pillow>=9.3.0
scikit_image==0.19.3
scikit_video==1.1.11
scipy==1.5.4
soundfile==0.11.0



tqdm==4.64.1
tslearn==0.5.3.2
omegaconf
pytorch-fid
git+https://github.com/facebookresearch/pytorch3d
